Add agent package and integrate into app
Introduce a new @jacoblincool/agent package (packages/agent) containing the agent runtime, dynamic sync, snapshot store, site/tool registries, clients, types and tests, and relocate static assets into the package. Replace many local implementations in apps/jacoblin.cool to consume agent exports (runtime, snapshot store, site registry, dynamic target resolution, builders, and tool definitions), and adapt function/schema shapes to Gemini-compatible declarations. Update app package.json to add a workspace dependency on @jacoblincool/agent and to build the agent before dev/build/check/test/preview commands. Update pnpm lockfile accordingly.
