+++ apps/jacoblin.cool/package.json
@@ -4,14 +4,14 @@
     "version": "0.0.1",
     "type": "module",
     "scripts": {
-        "dev": "firebase emulators:exec --ui 'vite dev'",
-        "build": "vite build",
-        "preview": "vite preview",
+        "dev": "pnpm --filter @jacoblincool/agent build && firebase emulators:exec --ui 'vite dev'",
+        "build": "pnpm --filter @jacoblincool/agent build && vite build",
+        "preview": "pnpm --filter @jacoblincool/agent build && vite preview",
         "prepare": "svelte-kit sync || echo ''",
-        "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+        "check": "pnpm --filter @jacoblincool/agent build && svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+        "check:watch": "pnpm --filter @jacoblincool/agent build && svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
         "lint": "eslint .",
-        "test": "vitest run",
+        "test": "pnpm --filter @jacoblincool/agent build && vitest run",
         "test:watch": "vitest"
     },
     "devDependencies": {
@@ -49,6 +49,7 @@
         "zod": "^4.3.6"
     },
     "dependencies": {
+        "@jacoblincool/agent": "workspace:*",
         "fires2rest": "^0.3.0",
         "jose": "^6.1.3",
         "marked": "^17.0.4",
