+++ packages/agent/assets/knowledge.graph.json
