+++ packages/agent/assets/knowledge.items.json
