+++ packages/agent/package.json
@@ -1,0 +1,25 @@
+{
+    "name": "@jacoblincool/agent",
+    "private": true,
+    "version": "0.0.1",
+    "type": "module",
+    "scripts": {
+        "build": "tsdown",
+        "test": "vitest run"
+    },
+    "devDependencies": {
+        "@types/node": "^25.2.3",
+        "tsdown": "^0.21.4",
+        "typescript": "^5.9.3",
+        "vitest": "^4.1.0"
+    },
+    "dependencies": {
+        "octokit": "^5.0.5",
+        "zod": "^4.3.6"
+    },
+    "exports": {
+        ".": "./scripts/index.mjs",
+        "./cli": "./scripts/cli.mjs",
+        "./package.json": "./package.json"
+    }
+}
