+++ packages/agent/tsconfig.json
@@ -1,0 +1,16 @@
+{
+    "compilerOptions": {
+        "target": "ES2022",
+        "module": "ESNext",
+        "moduleResolution": "bundler",
+        "strict": true,
+        "resolveJsonModule": true,
+        "skipLibCheck": true,
+        "esModuleInterop": true,
+        "isolatedModules": true,
+        "declaration": true,
+        "rewriteRelativeImportExtensions": true,
+        "types": ["node"]
+    },
+    "include": ["src/**/*.ts"]
+}
